// src/types/CSL.ts
export enum ShotSize {
  ExtremeCloseUp = "extremeCloseUp",
  CloseUp = "closeUp",
  MediumCloseUp = "mediumCloseUp",
  MediumShot = "mediumShot",
  MediumLongShot = "mediumLongShot",
  FullShot = "fullShot",
  LongShot = "longShot",
  VeryLongShot = "veryLongShot",
  ExtremeLongShot = "extremeLongShot",
}


export enum SubjectView {
  Front = "front",
  Back = "back",
  Left = "left",
  Right = "right",
  ThreeQuarterFrontLeft = "threeQuarterFrontLeft",
  ThreeQuarterFrontRight = "threeQuarterFrontRight",
  ThreeQuarterBackLeft = "threeQuarterBackLeft",
  ThreeQuarterBackRight = "threeQuarterBackRight",
}
// ...
export interface LossFunction {
  type: LossFunctionType;
  parameters: Record<string, unknown>;
}

export enum LossFunctionType {
  PedestalMovement = "pedestalMovement",
  DollyMovement = "dollyMovement",
  TruckMovement = "truckMovement",
  PanMovement = "panMovement",
  TiltMovement = "tiltMovement",
  ArcMovement = "arcMovement",
  Static = "Static",
  FollowMovement = "followMovement",

  FramingPosition = "framingPosition",
  ShotSize = "shotSize",
  SubjectView = "subjectView",

  // General losses
  Collision = "collision",
  Smoothness = "smoothness",
  MinPath = "minPath",
}

export interface SectionSolverOutput {
  initKeyframes: SinglePointConstraint[];
  constraints: (SinglePointConstraint | IntervalConstraint)[];
}

export interface TimelineSolverOutput {
  sections: SectionSolverOutput[];
}

export type TimelineSegment =
  | {
    kind: "interval";
    startTime: number;
    endTime: number;
    lossFunctions: LossFunction[];
  }
  | {
    kind: "point";
    time: number;
    lossFunctions: LossFunction[];
  };

export type NormalizedTimeline = TimelineSegment[];